Hi,

We have implemented the following project requirements,

### Project requirements 

- The following queries are taking too much time (**1 sec = 1000ms can be considered as too much time for database query**). Try them on your machine to confirm it:
  - `SELECT COUNT(*) FROM visits where animal_id = 4;`
  - `SELECT * FROM visits where vet_id = 2;`
  - `SELECT * FROM owners where email = 'owner_18327@mail.com';`
- Use `EXPLAIN ANALYZE` on the previous queries to check what is happening. Take screenshots of them - they will be necessary later.
- Find a way to decrease the execution time of the first query. Look for hints in the previous lessons.
- Find a way to improve execution time of the other two queries.
- While you are making changes, check if they help by running `EXPLAIN ANALYZE`. Once you succeed, take a screenshot of the `EXPLAIN ANALYZE` result showing that time actually decreased.
- Any changes you made to your database schema should be added to the schema.sql file and commited in a new pull request.
- If you decide to commit any `INSERT INTO` queries - remember to add the to the data.sql file. It is important to keep only the queries that change the database structure in the schema.sql file.
- In your pull request description include screenshots with `EXPLAIN ANALYZE` results (before and after you improve the speed) for each of the 3 problematic queries.


**OUTPUT**
```
-- Before optimization

vet_clinic=# explain analyze SELECT COUNT(*) FROM visits where animal_id = 4;
                                                                 QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=232172.60..232172.61 rows=1 width=8) (actual time=759.853..774.792 rows=1 loops=1)
   ->  Gather  (cost=232172.39..232172.60 rows=2 width=8) (actual time=759.766..774.787 rows=3 loops=1)
         Workers Planned: 2
         Workers Launched: 2
         ->  Partial Aggregate  (cost=231172.39..231172.40 rows=1 width=8) (actual time=717.517..717.518 rows=1 loops=3)
               ->  Parallel Seq Scan on visits  (cost=0.00..228891.55 rows=912334 width=0) (actual time=0.359..692.569 rows=718856 loops=3)
                     Filter: (animal_id = 4)
                     Rows Removed by Filter: 6469710
 Planning Time: 0.134 ms
 Execution Time: 774.823 ms
(10 rows)


vet_clinic=# explain analyze SELECT * FROM visits where vet_id = 2;
                                                       QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------
 Gather  (cost=1000.00..229893.45 rows=1 width=12) (actual time=601.534..618.973 rows=0 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on visits  (cost=0.00..228893.35 rows=1 width=12) (actual time=578.589..578.589 rows=0 loops=3)
         Filter: (vet_id = 2)
         Rows Removed by Filter: 7188567
 Planning Time: 0.147 ms
 Execution Time: 619.016 ms
(8 rows)


vet_clinic=# explain analyze SELECT * FROM owners where email = 'owner_18327@mail.com';
                                                        QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------
 Gather  (cost=1000.00..213276.03 rows=6 width=43) (actual time=4.699..1477.823 rows=6 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on owners  (cost=0.00..212275.43 rows=2 width=43) (actual time=267.904..1438.120 rows=2 loops=3)
         Filter: ((email)::text = 'owner_18327@mail.com'::text)
         Rows Removed by Filter: 5000000
 Planning Time: 6.678 ms
 Execution Time: 1477.840 ms
(8 rows)

-- After optimization

explain analyze SELECT COUNT(*) FROM visits where animal_id = 4;
                                                                            QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=190132.39..190132.40 rows=1 width=8) (actual time=251.208..259.475 rows=1 loops=1)
   ->  Gather  (cost=190132.18..190132.39 rows=2 width=8) (actual time=250.870..259.466 rows=3 loops=1)
         Workers Planned: 2
         Workers Launched: 2
         ->  Partial Aggregate  (cost=189132.18..189132.19 rows=1 width=8) (actual time=229.674..229.674 rows=1 loops=3)
               ->  Parallel Index Only Scan using animal_id on visits  (cost=0.44..186851.30 rows=912349 width=0) (actual time=0.191..205.108 rows=718856 loops=3)
                     Index Cond: (animal_id = 4)
                     Heap Fetches: 718860
 Planning Time: 0.148 ms
 Execution Time: 259.530 ms
(10 rows)

vet_clinic=# explain analyze SELECT * FROM visits where vet_id = 2;
                                                   QUERY PLAN
----------------------------------------------------------------------------------------------------------------
 Index Scan using vet_id on visits  (cost=0.44..8.21 rows=1 width=12) (actual time=0.011..0.011 rows=0 loops=1)
   Index Cond: (vet_id = 2)
 Planning Time: 0.090 ms
 Execution Time: 0.022 ms
(4 rows)


vet_clinic=# explain analyze SELECT * FROM owners where email = 'owner_18327@mail.com';
                                                      QUERY PLAN
----------------------------------------------------------------------------------------------------------------------
 Index Scan using owner_email on owners  (cost=0.56..28.62 rows=6 width=43) (actual time=0.045..0.055 rows=6 loops=1)
   Index Cond: ((email)::text = 'owner_18327@mail.com'::text)
 Planning Time: 0.136 ms
 Execution Time: 0.073 ms
(4 rows)
```